added basic login, signup and profile pages, added guest account, added correct, incorrect, and skip buttons for study page, updated css

+++ app.py
@@ -1,5 +1,5 @@
 # import libraries
-from flask import Flask, g, redirect, render_template, request, url_for
+from flask import Flask, g, redirect, render_template, request, url_for, session, flash
 # from werkzeug.utils import secure_filename  # for file uploads
 from datetime import datetime, timezone  # for time formatting
 import sqlite3
@@ -8,6 +8,9 @@ DATABASE = 'database.db'  # relative path to the database file
 
 # initialise app
 app = Flask(__name__)
+
+# set a secret key for sessions
+app.config['SECRET_KEY'] = "SecretKey123"
 
 
 # connect to the database
@@ -69,10 +72,24 @@ def format_date(date_string):
     return local_dt.strftime("%d/%b/%Y")
 
 
+# get user ID from session username
+def userID():
+    return session.get('userID', 0)
+
+
+# make username and userID available in all templates
+@app.context_processor
+def uservar():
+    return {
+        "username": session.get("username", "Guest"),
+        "userID": session.get("userID", 0)
+    }
+
+
 # homepage
 @app.route('/')
 def home():
-    return render_template("homepage.html")
+    return render_template("homepage.html", username=session.get('username', 'Guest'), userID=session.get('userID', 0))
 
 
 # list all decks
@@ -81,9 +98,10 @@ def Decks():
     # get all the decks id, name, description, and creation date
     sql = """
             SELECT deck_ID, deck_name, deck_description, deck_creation
-            FROM Decks;
+            FROM Decks
+            Where deck_userID = ?;
         """
-    result = query_db(sql)
+    result = query_db(sql, (userID(),))
     # return the results
     return render_template("decks.html", results=result)
 
@@ -95,15 +113,15 @@ def Deck(id):
     sql = """
             SELECT card_ID, card_question, card_answer, card_creation
             FROM Flashcards
-            WHERE card_deckID = ?;
+            WHERE card_deckID = ? AND card_userID = ?;
         """
     # get the deck name of the inputted deck id
     sql_deck = """
-                SELECT deck_name, deck_ID, deck_creation FROM Decks WHERE deck_ID = ?;
+                SELECT deck_name, deck_ID, deck_creation FROM Decks WHERE deck_ID = ? AND deck_userID = ?;
             """
 
-    deck_info = query_db(sql_deck, (id,), True)
-    results = query_db(sql, (id,))
+    deck_info = query_db(sql_deck, (id, userID()), True)
+    results = query_db(sql, (id, userID()))
     # return the results
     return render_template("deck.html", results=results, deck_info=deck_info, time_ago=time_ago, format_date=format_date)
 
@@ -112,9 +130,9 @@ def Deck(id):
 def deleteCard(id, card_id):
     # delete the card with the inputted card id
     sql = """
-            DELETE FROM Flashcards WHERE card_ID = ?;
+            DELETE FROM Flashcards WHERE card_ID = ? AND card_userID = ?;
         """
-    get_db().execute(sql, (card_id,))
+    get_db().execute(sql, (card_id, userID()))
     get_db().commit()
     # redirect to the deck page
     return redirect(url_for('Deck', id=id))
@@ -124,9 +142,9 @@ def deleteCard(id, card_id):
 def deleteDeck(id):
     # delete the deck with the inputted deck id
     sql = """
-            DELETE FROM Decks WHERE deck_ID = ?;
+            DELETE FROM Decks WHERE deck_ID = ? AND deck_userID = ?;
         """
-    get_db().execute(sql, (id,))
+    get_db().execute(sql, (id, userID()))
     get_db().commit()
     # redirect to the deck page
     return redirect(url_for('Decks'))
@@ -139,9 +157,9 @@ def Study(id, index):
     sql = """
             SELECT card_ID, card_question, card_answer, card_creation, card_hint
             FROM Flashcards
-            WHERE card_deckID = ?;
+            WHERE card_deckID = ? AND card_userID = ?;
         """
-    results = query_db(sql, (id,))
+    results = query_db(sql, (id, userID()))
 
     total = len(results)  # total number of cards in the deck
 
@@ -168,14 +186,15 @@ def createDeck():
         if not deck_name:
             # reload the page with the error message
             error = "A deck name is required."
+            flash("A deck name is required.")
             return render_template("deckCreate.html", error=error)
 
         else:
             sql = """
-                    INSERT INTO Decks (deck_name, deck_description, deck_creation)
-                    VALUES (?, ?, datetime('now'));
+                    INSERT INTO Decks (deck_name, deck_description, deck_creation, deck_userID)
+                    VALUES (?, ?, datetime('now'), ?);
                 """
-            get_db().execute(sql, (deck_name, deck_description))
+            get_db().execute(sql, (deck_name, deck_description, userID()))
             get_db().commit()
             # redirect to the decks list page
             return redirect(url_for('Decks'))
@@ -184,9 +203,10 @@ def createDeck():
         # get all the decks id, name, description, and creation date
         sql = """
                 SELECT deck_ID, deck_name, deck_description, deck_creation
-                FROM Decks;
+                FROM Decks
+                WHERE deck_userID = ?;
             """
-        result = query_db(sql)
+        result = query_db(sql, (userID(),))
         # return the results
         return render_template("deckCreate.html", results=result)
 
@@ -197,9 +217,9 @@ def createCard(id):
 
     # get the deck name of the inputted deck id
     sql_deck = """
-            SELECT deck_name, deck_ID, deck_creation FROM Decks WHERE deck_ID = ?;
+            SELECT deck_name, deck_ID, deck_creation FROM Decks WHERE deck_ID = ? AND deck_userID = ?;
         """
-    deck_info = query_db(sql_deck, (id,))[0]
+    deck_info = query_db(sql_deck, (id, userID()))[0]
 
     # if request method is POST, get the form data and insert into database
     if request.method == "POST":
@@ -211,16 +231,16 @@ def createCard(id):
         if not card_question or not card_answer:
             # reload the page with the error message
             error = "Both fields are required."
-
+            flash("Both fields are required.")
             return render_template("cardCreate.html", error=error, deck_info=deck_info)
 
         else:
             sql = """
-                    INSERT INTO Flashcards (card_question, card_answer, card_deckID, card_creation, card_hint)
-                    VALUES (?, ?, ?, datetime('now'), ?);
+                    INSERT INTO Flashcards (card_question, card_answer, card_deckID, card_creation, card_hint, card_userID)
+                    VALUES (?, ?, ?, datetime('now'), ?, ?);
                 """
 
-            get_db().execute(sql, (card_question, card_answer, id, card_hint))
+            get_db().execute(sql, (card_question, card_answer, id, card_hint, userID()))
             get_db().commit()
             # redirect to the deck page
             return redirect(url_for('Deck', id=id))
@@ -244,15 +264,17 @@ def editDeck(id):
         if not deck_name:
             # reload the page with the error message
             error = "A deck name is required."
+            flash("A deck name is required.")
             return render_template("deck.html", error=error, id=id)
 
         else:
             sql = """
                     UPDATE Decks
                     SET deck_name = ?, deck_description = ?
-                    WHERE deck_ID = ?;
+                    WHERE deck_ID = ?
+                    AND deck_userID = ?;
                 """
-            get_db().execute(sql, (deck_name, deck_description, id))
+            get_db().execute(sql, (deck_name, deck_description, id, userID()))
             get_db().commit()
             # redirect to the decks list page
             return redirect(url_for('Decks'))
@@ -262,9 +284,9 @@ def editDeck(id):
         sql = """
                 SELECT deck_ID, deck_name, deck_description, deck_creation
                 FROM Decks
-                WHERE deck_ID = ?;
+                WHERE deck_ID = ? AND deck_userID = ?;
             """
-        result = query_db(sql, (id,))
+        result = query_db(sql, (id, userID()))
         # return the results
         return render_template("deckEdit.html", results=result)
 
@@ -274,9 +296,9 @@ def editDeck(id):
 def editCard(id, card_id):
     # get the deck name of the inputted deck id
     sql_deck = """
-            SELECT deck_name, deck_ID, deck_creation FROM Decks WHERE deck_ID = ?;
+            SELECT deck_name, deck_ID, deck_creation FROM Decks WHERE deck_ID = ? AND deck_userID = ?;
         """
-    deck_info = query_db(sql_deck, (id,))[0]
+    deck_info = query_db(sql_deck, (id, userID()))[0]
 
     # if request method is POST, get the form data and insert into database
     if request.method == "POST":
@@ -288,19 +310,20 @@ def editCard(id, card_id):
         if not card_question or not card_answer:
             # reload the page with the error message
             error = "Both fields are required."
-
-            sql_card = "SELECT card_ID, card_question, card_answer, card_creation, card_hint FROM Flashcards WHERE card_ID = ?;"
-            card = query_db(sql_card, (card_id,), one=True)
+            flash("Both fields are required.")
+            sql_card = "SELECT card_ID, card_question, card_answer, card_creation, card_hint FROM Flashcards WHERE card_ID = ? AND card_userID = ?;"
+            card = query_db(sql_card, (card_id, userID()), one=True)
             return render_template("cardEdit.html", error=error, deck_info=deck_info, cards=card)
 
         else:
             sql = """
                     UPDATE Flashcards
                     SET card_question = ?, card_answer = ?, card_hint = ?
-                    WHERE card_ID = ? AND card_deckID = ?;
+                    WHERE card_ID = ? AND card_deckID = ?
+                    AND card_userID = ?;
                 """
 
-            get_db().execute(sql, (card_question, card_answer, card_hint, card_id, id))
+            get_db().execute(sql, (card_question, card_answer, card_hint, card_id, id, userID()))
             get_db().commit()
             # redirect to the deck page
             return redirect(url_for('Deck', id=id))
@@ -309,8 +332,8 @@ def editCard(id, card_id):
     else:
 
         # return cardinfo
-        sql_card = "SELECT card_ID, card_question, card_answer, card_creation, card_hint FROM Flashcards WHERE card_ID = ?;"
-        card = query_db(sql_card, (card_id,), one=True)
+        sql_card = "SELECT card_ID, card_question, card_answer, card_creation, card_hint FROM Flashcards WHERE card_ID = ? AND card_userID = ?;"
+        card = query_db(sql_card, (card_id, userID()), one=True)
         return render_template("cardEdit.html", deck_info=deck_info, cards=card)
 
 
@@ -321,6 +344,7 @@ def login():
         username = request.form['username']
         password = request.form['password']
 
+        # get usernames in the database
         sql = """
                 SELECT user_name, user_password
                 FROM Users
@@ -331,20 +355,25 @@ def login():
         # check if the username and password are not empty
         if not username or not password:
             error = "Both fields are required."
+            flash("Both fields are required.")
             return render_template("login.html", error=error)
 
         elif not username_list:
             # reload the page with the error message
             error = "No account under the username " + username + "."
+            flash("No account under the username " + username + ".")
             return render_template("login.html", error=error)
 
         else:
             # check if the password is correct
             if password == username_list[0][1]:
                 # logged in successfully, redirect to homepage
+                session['username'] = username
+                session['userID'] = query_db("SELECT user_ID FROM Users WHERE user_name = ?", (username,), one=True)[0]
                 return redirect(url_for('home'))
             else:
                 error = "Incorrect password."
+                flash("Incorrect password.")
                 return render_template("login.html", error=error)
 
     else:
@@ -357,6 +386,7 @@ def signup():
         # get the form data
         username = request.form['username']
         password = request.form['password']
+        confirm_password = request.form['confirm_password']
 
         sql = """
                 SELECT user_name
@@ -366,22 +396,57 @@ def signup():
         username_list = query_db(sql, (username,))
 
         # check if the username and password are not empty
-        if not username or not password:
-            error = "Both fields are required."
+        if not username or not password or not confirm_password:
+            error = "All fields are required."
+            flash("All fields are required.")
+            return render_template("signup.html", error=error)
+
+        elif confirm_password != password:
+            error = "Passwords do not match."
+            flash("Passwords do not match.")
             return render_template("signup.html", error=error)
 
         # check if the username already exists in the database
         elif username_list:
             error = "Username already exists."
+            flash("Username already exists.")
             return render_template("signup.html", error=error)
 
         else:
-            # reload the page with the error message
-            error = "Not accepting new signups yet."
-            return render_template("signup.html", error=error)
+            sql = """
+                    INSERT INTO Users (user_name, user_password, user_creation)
+                    VALUES (?, ?, datetime('now'));
+                """
+            get_db().execute(sql, (username, password))
+            get_db().commit()
+            error = "Account created successfully. Please log in."
+            flash("Account created successfully. Please log in.")
+            return render_template("login.html", error=error)
 
     else:
         return render_template("signup.html")
+
+
+@app.route('/profile/')
+def profile():
+    sql = """
+            SELECT user_name, user_creation
+            FROM Users
+            WHERE user_ID = ?;
+        """
+
+    results = query_db(sql, (userID(),))[0]
+    # return the results
+    return render_template("profile.html", results=results, format_date=format_date, time_ago=time_ago)
+
+
+@app.route('/logout/')
+def logout():
+    session.pop('username', None)
+    session.pop('userID', None)
+    flash("Logged out successfully.")
+    return redirect(url_for('home'))
+
 
 # only run the app if app.py is executed directly
 if __name__ == "__main__":
